// src/components/helpers/types.ts
export enum BUTTON_SIZE {
    MINI = 'mini',
    SMALL = 'small',
    DEFAULT = 'medium',
    LARGE = 'large'
}

export enum BUTTON_VARIANTS {
    DEFAULT = "default",
    PRIMARY = "primary",
    SECONDARY = "secondary",
    SUCCESS = "success",
    WARNING = "warning",
    DANGER = "danger",
}

export enum BUTTON_ROUNDED {
    NONE = '',
    BOTH = 'rounded',
    LEFT = 'left-rounded',
    RIGHT = 'right-rounded'
}

export enum BUTTON_ANIMATION {
    NONE = '',
    ROTATE = 'rotate-icon',
    FLASH = 'flash-icon',
    HEARTBEAT = 'heartbeat-icon',
    RING = 'ring-icon',

}

export enum BUTTON_EXTRA {
    NONE = '',
    SQUARE = 'square-button',
    CIRCLE = 'circle-button',
}